+++ src/threads.js
@@ -61,7 +61,7 @@ StageMorph, SpriteMorph, StagePrompterMorph, Note, modules, isString, copy, Map,
 isNil, WatcherMorph, List, ListWatcherMorph, alert, console, TableMorph, BLACK,
 TableFrameMorph, ColorSlotMorph, isSnapObject, newCanvas, Symbol, SVG_Costume*/
 
-modules.threads = '2020-November-15';
+modules.threads = '2020-November-20';
 
 var ThreadManager;
 var Process;
@@ -4904,8 +4904,9 @@ Process.prototype.reportGet = function (query) {
             return new List(
                 stage.children.filter(each =>
                     each instanceof SpriteMorph &&
+                        each.isVisible &&
                         (each !== thisObj) &&
-                            each.bounds.intersects(neighborhood)
+                        each.bounds.intersects(neighborhood)
                 )
             );
         case 'dangling?':