+++ HISTORY.md
@@ -4,6 +4,10 @@
 
 * **Notable Fixes:**
     * fixed stretching SVG costumes with fixed aspect ratios in Firefox
+    * only report neighbors that are visible
+
+### 2020-11-20
+* threads: only report neighbors that are visible, thanks Frederic, for reporting this bug!
 
 ### 2020-11-19
 * new dev version